Crate diff

Enums

Result

A fragment of a computed diff.

Functions

chars

Computes the diff between the chars of two strings.

lines

Computes the diff between the lines of two strings.

slice

Computes the diff between two slices.