Enum diff::DiffItem

pub enum DiffItem<'a, T: 'a> where
    T: PartialEq + Debug,  {
    Add {
        start_from: usize,
        start_to: usize,
        end_to: usize,
        items: &'a [T],
    },
    Delete {
        start_from: usize,
        end_from: usize,
        start_to: usize,
        items: &'a [T],
    },
    Change {
        start_from: usize,
        end_from: usize,
        start_to: usize,
        end_to: usize,
        from: &'a [T],
        to: &'a [T],
    },
}

The DiffItem enum. Represents an edit action (either Add, Delete, or Change) Contains the necessary information to make the corresponding change to a slice.

Variants
Add

Represents an insertion edit

Fields of Add

start_from: usize

The index at which to insert the new items in the "from" slice

start_to: usize

The starting index of the new items in the "to" slice

end_to: usize

The ending index of the new items in the "to" slice

items: &'a [T]

The items to be inserted

Delete

Represents a deletion edit

Fields of Delete

start_from: usize

The starting index of the items to delete from the "from" slice

end_from: usize

The ending index of the items to delete from the "from" slice

start_to: usize

The index in "to" that corresponds to "start_from"

items: &'a [T]

The items to be deleted

Change

Represents a change edit

Fields of Change

start_from: usize

The starting index of the items to Change from the "from" slice

end_from: usize

The ending index of the items to Change from the "from" slice

start_to: usize

The starting index of the items to Change from the "to" slice

end_to: usize

The ending index of the items to Change from the "to" slice

from: &'a [T]

The lines that will be Changed in "from"

to: &'a [T]

The change lines from "to"